Apply review comments
