+++ jetstream/jetstream_options.go
@@ -1,4 +1,4 @@
-// Copyright 2022-2023 The NATS Authors
+// Copyright 2022-2024 The NATS Authors
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -309,7 +309,9 @@ func WithMsgID(id string) PublishOpt {
 	}
 }
 
-// WithExpectStream sets the expected stream to respond from the publish.
+// WithExpectStream sets the expected stream the message should be published to.
+// If the message is published to a different stream server will reject the
+// message and publish will fail.
 func WithExpectStream(stream string) PublishOpt {
 	return func(opts *pubOpts) error {
 		opts.stream = stream
@@ -317,8 +319,9 @@ func WithExpectStream(stream string) PublishOpt {
 	}
 }
 
-// WithExpectLastSequence sets the expected sequence in the response from the
-// publish.
+// WithExpectLastSequence sets the expected sequence number the last message
+// on a stream should have. If the last message has a different sequence number
+// server will reject the message and publish will fail.
 func WithExpectLastSequence(seq uint64) PublishOpt {
 	return func(opts *pubOpts) error {
 		opts.lastSeq = &seq
@@ -326,8 +329,10 @@ func WithExpectLastSequence(seq uint64) PublishOpt {
 	}
 }
 
-// WithExpectLastSequencePerSubject sets the expected sequence per subject in
-// the response from the publish.
+// WithExpectLastSequencePerSubject sets the expected sequence number the last
+// message on a subject the message is published to. If the last message on a
+// subject has a different sequence number server will reject the message and
+// publish will fail.
 func WithExpectLastSequencePerSubject(seq uint64) PublishOpt {
 	return func(opts *pubOpts) error {
 		opts.lastSubjectSeq = &seq
@@ -335,8 +340,9 @@ func WithExpectLastSequencePerSubject(seq uint64) PublishOpt {
 	}
 }
 
-// WithExpectLastMsgID sets the expected last msgId in the response from the
-// publish.
+// WithExpectLastMsgID sets the expected message ID the last message on a stream
+// should have. If the last message has a different message ID server will
+// reject the message and publish will fail.
 func WithExpectLastMsgID(id string) PublishOpt {
 	return func(opts *pubOpts) error {
 		opts.lastMsgID = id
@@ -377,118 +383,6 @@ func WithStallWait(ttl time.Duration) PublishOpt {
 			return fmt.Errorf("%w: stall wait should be more than 0", ErrInvalidOption)
 		}
 		opts.stallWait = ttl
-		return nil
-	}
-}
-
-// KV Options
-
-type watchOptFn func(opts *watchOpts) error
-
-func (opt watchOptFn) configureWatcher(opts *watchOpts) error {
-	return opt(opts)
-}
-
-// IncludeHistory instructs the key watcher to include historical values as
-// well.
-func IncludeHistory() WatchOpt {
-	return watchOptFn(func(opts *watchOpts) error {
-		if opts.updatesOnly {
-			return fmt.Errorf("%w: include history can not be used with updates only", ErrInvalidOption)
-		}
-		opts.includeHistory = true
-		return nil
-	})
-}
-
-// UpdatesOnly instructs the key watcher to only include updates on values
-// (without latest values when started).
-func UpdatesOnly() WatchOpt {
-	return watchOptFn(func(opts *watchOpts) error {
-		if opts.includeHistory {
-			return fmt.Errorf("%w: updates only can not be used with include history", ErrInvalidOption)
-		}
-		opts.updatesOnly = true
-		return nil
-	})
-}
-
-// IgnoreDeletes will have the key watcher not pass any deleted keys.
-func IgnoreDeletes() WatchOpt {
-	return watchOptFn(func(opts *watchOpts) error {
-		opts.ignoreDeletes = true
-		return nil
-	})
-}
-
-// MetaOnly instructs the key watcher to retrieve only the entry meta data, not
-// the entry value.
-func MetaOnly() WatchOpt {
-	return watchOptFn(func(opts *watchOpts) error {
-		opts.metaOnly = true
-		return nil
-	})
-}
-
-// ResumeFromRevision instructs the key watcher to resume from a specific
-// revision number.
-func ResumeFromRevision(revision uint64) WatchOpt {
-	return watchOptFn(func(opts *watchOpts) error {
-		opts.resumeFromRevision = revision
-		return nil
-	})
-}
-
-// DeleteMarkersOlderThan indicates that delete or purge markers older than that
-// will be deleted as part of PurgeDeletes() operation, otherwise, only the data
-// will be removed but markers that are recent will be kept.
-// Note that if no option is specified, the default is 30 minutes. You can set
-// this option to a negative value to instruct to always remove the markers,
-// regardless of their age.
-type DeleteMarkersOlderThan time.Duration
-
-func (ttl DeleteMarkersOlderThan) configurePurge(opts *purgeOpts) error {
-	opts.dmthr = time.Duration(ttl)
-	return nil
-}
-
-type deleteOptFn func(opts *deleteOpts) error
-
-func (opt deleteOptFn) configureDelete(opts *deleteOpts) error {
-	return opt(opts)
-}
-
-// LastRevision deletes if the latest revision matches.
-func LastRevision(revision uint64) KVDeleteOpt {
-	return deleteOptFn(func(opts *deleteOpts) error {
-		opts.revision = revision
-		return nil
-	})
-}
-
-// ObjectStore options
-
-// GetObjectShowDeleted makes Get() return object if it was marked as deleted.
-func GetObjectShowDeleted() GetObjectOpt {
-	return func(opts *getObjectOpts) error {
-		opts.showDeleted = true
-		return nil
-	}
-}
-
-// GetObjectInfoShowDeleted makes GetInfo() return object if it was marked as
-// deleted.
-func GetObjectInfoShowDeleted() GetObjectInfoOpt {
-	return func(opts *getObjectInfoOpts) error {
-		opts.showDeleted = true
-		return nil
-	}
-}
-
-// ListObjectsShowDeleted makes ListObjects() return deleted objects.
-func ListObjectsShowDeleted() ListObjectsOpt {
-	return func(opts *listObjectOpts) error {
-		opts.showDeleted = true
 		return nil
 	}
 }
